remove uncessary accessibility component screen
